Complete Stage 2 (Connection)
Add three more components (combo box, text area and another button) to the window, and implement limited functionality.

On clicking the "Open" button, take the text in the text field and try to open a database with the specified name. If successful, read in the table names and add them to the combo box.

Display a "SELECT" query in the text area to read all records from the table selected in the combo box.